% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{graph}
\alias{graph}
\title{Graph constructor}
\usage{
graph(nodes, links)
}
\arguments{
\item{nodes}{A vector of integers (vertices).}

\item{links}{A list of vectors (edges).}
}
\value{
A list representing a graph.
}
\description{
This function is for constructing graphs.
}
