% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotgraph}
\alias{plotgraph}
\title{Plotting graphs}
\usage{
plotgraph(g, ...)
}
\arguments{
\item{g}{A list representing a graph.}

\item{...}{Aditional style arguments.}
}
\value{
Graph (invisibly).
}
\description{
This function is for plotting graphs.
}
